+++ terminaltexteffects/effects/effect_template.py
@@ -1,0 +1,193 @@
+"""Effect Description.
+
+Classes:
+
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+import terminaltexteffects as tte
+from terminaltexteffects.engine.base_config import BaseConfig
+from terminaltexteffects.engine.base_effect import BaseEffect, BaseEffectIterator
+from terminaltexteffects.engine.effect_support.particles import ParticlePool
+from terminaltexteffects.utils import argutils
+from terminaltexteffects.utils.argutils import ArgSpec, ParserSpec
+
+
+def get_effect_resources() -> tuple[str, type[BaseEffect], type[BaseConfig]]:
+    """Get the command, effect class, and configuration class for the effect.
+
+    Returns:
+        tuple[str, type[BaseEffect], type[BaseConfig]]: The command name, effect class, and configuration class.
+
+    """
+    return "dev", Effect, EffectConfig
+
+
+@dataclass
+class EffectConfig(BaseConfig):
+    """Effect configuration dataclass."""
+
+    parser_spec: ParserSpec = ParserSpec(
+        name="dev",
+        help="effect_description",
+        description="effect_description",
+        epilog=f"""{argutils.EASING_EPILOG}
+    """,
+    )
+
+    color_single: tte.Color = ArgSpec(
+        name="--color-single",
+        type=argutils.ColorArg.type_parser,
+        default=tte.Color(0),
+        metavar=argutils.ColorArg.METAVAR,
+        help="Color for the ___.",
+    )  # pyright: ignore[reportAssignmentType]
+    "Color: Color for the ___."
+
+    final_gradient_stops: tuple[tte.Color, ...] = ArgSpec(
+        name="--final-gradient-stops",
+        type=argutils.ColorArg.type_parser,
+        nargs="+",
+        action=argutils.TupleAction,
+        default=(tte.Color("#8A008A"), tte.Color("#00D1FF"), tte.Color("#FFFFFF")),
+        metavar=argutils.ColorArg.METAVAR,
+        help=(
+            "Space separated, unquoted, list of colors for the character gradient (applied across the canvas). "
+            "If only one color is provided, the characters will be displayed in that color."
+        ),
+    )  # pyright: ignore[reportAssignmentType]
+    (
+        "tuple[Color, ...]: Space separated, unquoted, list of colors for the character gradient "
+        "(applied across the canvas). If only one color is provided, the characters will be displayed in that color."
+    )
+
+    final_gradient_steps: tuple[int, ...] | int = ArgSpec(
+        name="--final-gradient-steps",
+        type=argutils.PositiveInt.type_parser,
+        nargs="+",
+        action=argutils.TupleAction,
+        default=12,
+        metavar=argutils.PositiveInt.METAVAR,
+        help=(
+            "Space separated, unquoted, list of the number of gradient steps to use. More steps will "
+            "create a smoother and longer gradient animation."
+        ),
+    )  # pyright: ignore[reportAssignmentType]
+    (
+        "tuple[int, ...] | int: Space separated, unquoted, list of the number of gradient steps to use. More "
+        "steps will create a smoother and longer gradient animation."
+    )
+
+    final_gradient_frames: int = ArgSpec(
+        name="--final-gradient-frames",
+        type=argutils.PositiveInt.type_parser,
+        default=5,
+        metavar=argutils.PositiveInt.METAVAR,
+        help="Number of frames to display each gradient step. Increase to slow down the gradient animation.",
+    )  # pyright: ignore[reportAssignmentType]
+    "int: Number of frames to display each gradient step. Increase to slow down the gradient animation."
+
+    final_gradient_direction: tte.Gradient.Direction = ArgSpec(
+        name="--final-gradient-direction",
+        type=argutils.GradientDirection.type_parser,
+        default=tte.Gradient.Direction.VERTICAL,
+        metavar=argutils.GradientDirection.METAVAR,
+        help="Direction of the final gradient.",
+    )  # pyright: ignore[reportAssignmentType]
+    "Gradient.Direction : Direction of the final gradient."
+
+    movement_speed: float = ArgSpec(
+        name="--movement-speed",
+        type=argutils.PositiveFloat.type_parser,
+        default=1,
+        metavar=argutils.PositiveFloat.METAVAR,
+        help="Speed of the ___.",
+    )  # pyright: ignore[reportAssignmentType]
+    "float: Speed of the ___."
+
+    easing: tte.easing.EasingFunction = ArgSpec(
+        name="--easing",
+        default=tte.easing.in_out_sine,
+        type=argutils.Ease.type_parser,
+        help="Easing function to use for character movement.",
+    )  # pyright: ignore[reportAssignmentType]
+    "easing.EasingFunction: Easing function to use for character movement."
+
+
+class EffectIterator(BaseEffectIterator[EffectConfig]):
+    """Effect iterator for the NamedEffect effect."""
+
+    def __init__(self, effect: Effect) -> None:
+        """Initialize the effect iterator.
+
+        Args:
+            effect (NamedEffect): The effect to iterate over.
+
+        """
+        super().__init__(effect)
+        self.pending_chars: list[tte.EffectCharacter] = []
+        self.character_final_color_map: dict[tte.EffectCharacter, tte.Color] = {}
+        self.spawner = self.terminal.add_character("#", tte.Coord(1, 1))
+        self.particle_gap = 10
+        self.frames_until_particle = 0
+        self.pool = self._setup_pool()
+        self.build()
+
+    def _setup_pool(self) -> ParticlePool:
+        def _setup_particle(particle_char: tte.EffectCharacter) -> None:
+            fade_scn = particle_char.animation.new_scene(scene_id="fade")
+            for color in tte.Gradient(tte.Color("ff0000"), tte.Color("000000"), steps=10):
+                fade_scn.add_frame("0", 4, colors=tte.ColorPair(fg=color))
+
+        pool = ParticlePool(
+            self.terminal,
+            self.active_characters,
+            symbols="O",
+            initial_count=1000,
+            initializer=_setup_particle,
+        )
+        for particle in pool.particles:
+            pool.reclaim_on_event(
+                particle,
+                caller=particle.animation.query_scene("fade"),
+            )
+        return pool
+
+    def build(self) -> None:
+        """Build the effect."""
+        p = self.spawner.motion.new_path(path_id="route", speed=0.8, ease=tte.easing.out_bounce)
+        for _ in range(10):
+            p.new_waypoint(self.terminal.canvas.random_coord(), bezier_control=self.terminal.canvas.random_coord())
+        self.spawner.motion.activate_path(p)
+        self.terminal.set_character_visibility(self.spawner, is_visible=True)
+        self.active_characters.add(self.spawner)
+
+    def __next__(self) -> str:
+        """Return the next frame of the effect."""
+        if self.pending_chars or self.active_characters:
+            if not self.frames_until_particle and self.spawner in self.active_characters:
+                self.pool.emit(
+                    origin=self.spawner.motion.current_coord,
+                    on_emit=lambda c: c.animation.activate_scene("fade"),
+                )
+                self.frames_until_particle = self.particle_gap
+            else:
+                self.frames_until_particle -= 1
+            self.update()
+            return self.frame
+        raise StopIteration
+
+
+class Effect(BaseEffect[EffectConfig]):
+    """Effect description."""
+
+    @property
+    def _config_cls(self) -> type[EffectConfig]:
+        return EffectConfig
+
+    @property
+    def _iterator_cls(self) -> type[EffectIterator]:
+        return EffectIterator